wykop/accounts/serializers.py
from rest_framework import serializers

from wykop.accounts.models import User


class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ('username', 'banned')
