requirements.txt
django
django-extensions
django-debug-toolbar
bpython
Pillow
django-embed-video
gunicorn
django-heroku
django-bootstrap-pagination
djangorestframework
